- replace old-style tendrl objects with new
tendrl-bug-id: Tendrl/specifications#37
tendrl-bug-id: Tendrl/specifications#54
tendrl-bug-id: Tendrl/specifications#32
