+++ tendrl/node_agent/objects/node_context/__init__.py
@@ -1,0 +1,68 @@
+import logging
+import os
+import socket
+import uuid
+
+from tendrl.commons.etcdobj.etcdobj import EtcdObj
+from tendrl.commons.utils import cmd_utils
+
+from tendrl.node_agent import objects
+
+
+LOG = logging.getLogger(__name__)
+
+
+class NodeContext(objects.NodeAgentBaseObject):
+    def __init__(self, machine_id=None, node_id=None, fqdn=None,
+                 tags=None, status=None, *args, **kwargs):
+        super(NodeContext, self).__init__(*args, **kwargs)
+
+        self.value = 'nodes/%s/NodeContext'
+        self.machine_id = machine_id or self._get_machine_id()
+        self.node_id = node_id or self._get_node_id() or self.create_node_id()
+        self.fqdn = fqdn or socket.getfqdn()
+        self.tags = tags
+        self.status = status or "UP"
+        self._etcd_cls = _NodeContextEtcd
+
+    def _get_machine_id(self):
+        cmd = cmd_utils.Command({"_raw_params": "cat /etc/machine-id"})
+        out, err, rc = cmd.run(tendrl_ns.config['tendrl_ansible_exec_file'])
+        return out['stdout']
+
+    def _create_node_id(self, node_id=None):
+        node_id = node_id or str(uuid.uuid4())
+        local_node_context = "~/.tendrl/" + self.value % self.node_id + \
+                             "/node_id"
+        with open(local_node_context, 'wb+') as f:
+            f.write(node_id)
+            LOG.info("SET_LOCAL: "
+                     "tendrl_ns.node_agent.objects.NodeContext.node_id==%s" %
+                     self.node_id)
+
+    def _get_node_id(self):
+        try:
+            local_node_context = "~/.tendrl/" + self.value % self.node_id + \
+                                 "/node_id"
+            if os.path.isfile(local_node_context):
+                with open(local_node_context) as f:
+                    node_id = f.read()
+                    if node_id:
+                        LOG.info(
+                            "GET_LOCAL: tendrl_ns.node_agent.objects.NodeContext"
+                            ".node_id==%s" % node_id)
+                        return node_id
+        except AttributeError:
+            return None
+
+
+class _NodeContextEtcd(EtcdObj):
+    """A table of the node context, lazily updated
+
+    """
+    __name__ = 'nodes/%s/Node_context'
+    _tendrl_cls = NodeContext
+
+    def render(self):
+        self.__name__ = self.__name__ % self.node_id
+        return super(_NodeContextEtcd, self).render()
